+++ src/console/controllers/ThumbsController.php
@@ -99,7 +99,7 @@ class ThumbsController extends Controller
             $meta = Aws::getInstance()->assets->createAllThumbSizes($asset);
 
             $cachedMeta = [
-                'baseUrl' => $baseUrl . '/' . $meta['path'],
+                'baseUrl' => $baseUrl . $meta['path'],
                 'ext' => $meta['ext'],
             ];