+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Craft AWS Plugin Changelog
 
-## 1.0.0-beta.2 - 2019-04-08
+## Unreleased
+### Added
+- Console command to generate thumbs without job/queues
 ### Fixed
-- If resourceRevision was a clousure, an error occured.
+- If resourceRevision was a clousure, an error occured
+- Fix AssetManager by adding S3 functions as trait
 
 ## 1.0.0-beta.1 - 2019-04-06
 ### Added