+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Fixed
 - If resourceRevision was a clousure, an error occured
 - Fix AssetManager by adding S3 functions as trait
+- Fix cached path of thumbnails
 
 ## 1.0.0-beta.1 - 2019-04-06
 ### Added