+++ CHANGELOG.md
@@ -1,14 +1,18 @@
 # Craft AWS Plugin Changelog
 
-## Unreleased
+## 1.0.0-beta.2 - 2020-03-01
+
 ### Added
 - Console command to generate thumbs without job/queues
 - Config option for queue name for all mass updates (e.g. thumbs generating)
+
 ### Fixed
 - If resourceRevision was a clousure, an error occured
 - Fix AssetManager by adding S3 functions as trait
 - Fix cached path of thumbnails
+- Fix SVG mime type for older Craft CMS versions
 
 ## 1.0.0-beta.1 - 2019-04-06
+
 ### Added
 - Initial release